+++ trunk/src/scf-base.cpp
@@ -183,7 +183,7 @@ SCF::SCF(const BasisSet & basis, const Settings & set, Checkpoint & chkpt) {
       
       if(verbose) {
 	N=tab.memory_estimate(&basis);
-	printf("Forming table of %lu ERIs, requiring %s of memory ... ",N,memory_size(N).c_str());
+	printf("Forming table of %lu ERIs, requiring %s of memory ... ",(long unsigned int) N,memory_size(N).c_str());
 	fflush(stdout);
       }
       tab.fill(&basis);
